# Use assembly routines for Dot and AddScaled*. Add benchmark for AddScale...