@btracey 2015-01-11
The change to AddScaledTo is not a clear win. 

brendan:~/Documents/mygo/src/github.com/gonum/floats$ benchcmp orig.bench new.bench 
benchmark                      old ns/op     new ns/op     delta       
BenchmarkAddScaledToSmall      20.0          21.2          +6.00%      
BenchmarkAddScaledToMedium     1249          525           -57.97%     
BenchmarkAddScaledToLarge      125598        105003        -16.40%     
BenchmarkAddScaledToHuge       21816250      27014749      +23.83%  

However, the change is minor for small entries, very significantly positive for medium sizes, and significantly negative for Huge. It seems to me this is a win because things will usually be smaller, and I've seen a number of PRs over time to improve Copy, and if/when we have AVX the assembly routine will be faster. Both of these should speed up Huge and make it a net win. If not, we could add a size length switch of some kind.