@btracey 2015-01-11
SGTM

If you would like to make that change,  I'll update this and the blas PRs. Thanks for the input,  and sorry if I'm being overly cautious.