@fhs 2015-01-12
What's the benchmark for AddScaledTo now? Did it improve?