@btracey 2015-01-12
Much better, thank you

benchmark                      old ns/op     new ns/op     delta       
BenchmarkAddScaledToSmall      20.9          14.0          -33.01%     
BenchmarkAddScaledToMedium     1219          318           -73.91%     
BenchmarkAddScaledToLarge      120523        70862         -41.20%     
BenchmarkAddScaledToHuge       19079937      19094861      +0.08%