@fhs 2015-01-12
LGTM

The benchmarks can be made more compact by calling RandomSlice from within benchmarkAddScaledTo but I'll leave it up to you.