@btracey 2015-01-12
I'm going to submit as is and make a PR to fix all the benchmarks.